refactor(renderer): extract SSR march + temporal denoiser into ssr_pass.rs
Render-graph migration cluster 2 prep: both SSR passes become
self-contained Renderer methods (the march keeps its internal
profiler spans; the denoiser needs only the encoder), trimmed to the
parameters their bodies actually use. Goldens pixel-identical.
mod.rs ratchets 13591 -> 13463.
