+++ tools/file-lines-baseline.json
@@ -1,3 +1,3 @@
 {
-  "native/shared/src/renderer/mod.rs": 13591
+  "native/shared/src/renderer/mod.rs": 13464
 }
